week8_query_optimization/clarke_csc6302_database_principles_project_mod8/clarkes_mbd_app/requirements.txt
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
flask-mysql-connector==1.1.0
Flask-WTF==1.2.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
narwhals==2.14.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
plotly==6.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
